requirements.txt
runpod
ultralytics
Pillow
requests